+++ src/components/Modal/Modal.tsx
@@ -1,0 +1,35 @@
+import React, { useEffect, useState, ReactNode } from "react";
+
+import ModalPortal from "./ModalPortal";
+
+import { useModal } from "../../hooks";
+import { ModalBlock, Dimmed, Contents } from "./Modal.styles";
+
+interface Props {
+  children: ReactNode;
+}
+
+const Modal = ({ children }: Props) => {
+  const [opacity, setOpacity] = useState(0);
+  const { close } = useModal();
+
+  useEffect(() => {
+    setOpacity(1);
+  }, []);
+
+  const dimmedClick: React.MouseEventHandler = ({ target, currentTarget }) => {
+    if (target !== currentTarget) return;
+
+    close();
+  };
+
+  return ModalPortal(
+    <ModalBlock>
+      <Dimmed opacity={opacity} onClick={dimmedClick} />
+      <Contents>{children}</Contents>
+    </ModalBlock>
+  );
+};
+
+export default Modal;
+export type { Props };
